Partway through ex 34 - NOT IN COMPILING STATE
It needs careful review of the edge cases and some documentation.

I'm taking a break from this project to start learning C++; I
definitely think it's still going all the way through Learn C
the Hard Way, but I feel like I've accomplished my original goal,
which was to make sure I mastered the basics of C before learning
C++, so that the boundary between the languages would be clear
to me if / when I need to drop down to C.

But ultimately as a quantitative researcher, I'm far more likely
to write code in C++. So it makes sense to shift my focus there
at this point.